Fix Build Errors (#1635)
* Fix build err when built with no thread

* Enable KAT

* Add Generated test

* Fix typo

* update formatting

* Fix typo

* Fix build issues

* Fix undefined error

* Fixed SA issue

* Fixed warnings

* Skip variant that exceeds timeout

* Fix style format

* Fix various build issues

+++ src/sig_stfl/lms/external/endian.c
@@ -4,7 +4,7 @@ void put_bigendian( void *target, unsigned long long value, size_t bytes ) {
     unsigned char *b = target;
     int i;
 
-    for (i = bytes-1; i >= 0; i--) {
+    for (i = (int)(bytes-1); i >= 0; i--) {
         b[i] = value & 0xff;
         value >>= 8;
     }
